move ncbi-calling funcs to ncbi mod
move the functions that perform the called to NCBI.Entrez to the NCBI module
